requirements.txt
fastapi
uvicorn
redis
transformers
python-telegram-bot
python-dotenv
tweepy